Implement htmx in login form and create auth route
The commit includes several significant changes. The login form in the FlaredUI Templates now uses htmx for dynamic content exchanges, improving the user experience. A new route for authentication was created, and CORS, previously enabled in the Flask application for testing, was disabled. This commit also introduces the ability to dynamically register blueprints, enhancing the flexibility of the application. A script tag was added to the base.html file to include htmx library in the project.